+++ test/asv.linux.conf.json
@@ -42,6 +42,7 @@
     ],
     "build_command": [
         "python -mpip install -U scikit-build cython",
+        "pip install git+https://github.com/Qiskit/qiskit-terra",
         "python setup.py bdist_wheel --dist-dir={build_cache_dir} -- -DCMAKE_CXX_COMPILER=g++ -- -j"
     ],
