+++ test/benchmark2/noise.py
@@ -1,14 +1,13 @@
 from benchmark2.simulator_benchmark import SimulatorBenchmarkSuite
-from benchmark2.particular_quantum_circuits import ParticularQuantumCircuits
 
-DEFAULT_APPS = [
+DEFAULT_APPS = {
 #    'fourier_checking',
 #    'graph_state',
 #    'hidden_linear_function',
 #    'iqp',
-    'quantum_volume'
+    'quantum_volume': 1
 #    'phase_estimation'
-    ]
+    }
 
 DEFAULT_QUBITS = [10, 11, 12, 13, 14]
 
@@ -30,7 +29,7 @@ DEFAULT_NOISE_MODELS = [
     SimulatorBenchmarkSuite.NOISE_DEPOLARIZING
     ]
 
-class NoiseParticularQuantumCircuits(ParticularQuantumCircuits):
+class DampingError(SimulatorBenchmarkSuite):
 
     def __init__(self,
                  apps = DEFAULT_APPS,
@@ -38,15 +37,34 @@ class NoiseParticularQuantumCircuits(ParticularQuantumCircuits):
                  runtime_names = DEFAULT_RUNTIME,
                  measures = DEFAULT_MEASUREMENT_METHODS,
                  measure_counts = DEFAULT_MEASUREMENT_COUNTS,
-                 noise_model_names = DEFAULT_NOISE_MODELS):
-        super().__init__( apps, 
-                          qubits=qubits, 
-                          runtime_names=runtime_names, 
-                          measures=measures, 
-                          measure_counts=measure_counts, 
-                          noise_model_names=noise_model_names)
-        self.__name__ = 'noise_particular_quantum_circuits'
+                 noise_model_names = [SimulatorBenchmarkSuite.NOISE_DAMPING] ):
+        super().__init__('depolarizing_error',
+                         apps, 
+                         qubits=qubits, 
+                         runtime_names=runtime_names, 
+                         measures=measures, 
+                         measure_counts=measure_counts, 
+                         noise_model_names=noise_model_names)
+        self.__name__ = 'damping_error'
         
+class DepolarizingError(SimulatorBenchmarkSuite):
+
+    def __init__(self,
+                 apps = DEFAULT_APPS,
+                 qubits = DEFAULT_QUBITS,
+                 runtime_names = DEFAULT_RUNTIME,
+                 measures = DEFAULT_MEASUREMENT_METHODS,
+                 measure_counts = DEFAULT_MEASUREMENT_COUNTS,
+                 noise_model_names = [SimulatorBenchmarkSuite.NOISE_DEPOLARIZING] ):
+        super().__init__('depolarizing_error',
+                         apps, 
+                         qubits=qubits, 
+                         runtime_names=runtime_names, 
+                         measures=measures, 
+                         measure_counts=measure_counts, 
+                         noise_model_names=noise_model_names)
+        self.__name__ = 'depolarizing_error'
     
 if __name__ == "__main__":
-    NoiseParticularQuantumCircuits().run_manual()
+    DampingError().run_manual()
+    DepolarizingError().run_manual()
